Detach the Audio source file from the app.cpp
